Huge update
Beaucoup de changements à lister, mais pour faire le plus court possible la plupart il reste les tâches suivantes:
- L'ajout au panier depuis la page des produits
- Mise à jour en temps réel du nombre de produits dans le panier dans la navbar
- Changements visuels dans la page de visualisation des produits
- Certains messages ne sont pas traduit
- Mettre à jour l'affichage du panier (légèrement)
- Rajouter un bouton pour supprimer des produits de son panier pour chaque produit
- Mettre à jour cart_view (controlleur)

+++ src/Controller/ProductController.php
@@ -7,25 +7,30 @@ use App\Entity\Product;
 use App\Form\ProductType;
 use App\Repository\ProductRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Translation\LocaleSwitcher;
 
 class ProductController extends AbstractController
 {
-    #[Route('/product/list.html.twig', name: 'product_list')]
-    public function list(ProductRepository $product, Request $request, LocaleSwitcher $localeSwitcher): Response
+    #[Route('/product/list', name: 'product_list')]
+    public function list(ProductRepository $product, Request $request, LocaleSwitcher $localeSwitcher, PaginatorInterface $paginator): Response
     {
         $locale = $request->getSession()->get('_locale', 'en');
         $localeSwitcher->setLocale($locale);
 
+        $query = $request->query->get('q', '');
+
         $products = $product->findAllProducts();
 
         return $this->render('product/list.html.twig', [
-            'products' => $products,
+            'pagination' => $products,
+            'searchQuery' => $query,
         ]);
     }
 
@@ -61,6 +66,8 @@ class ProductController extends AbstractController
             $entityManager->persist($product);
             $entityManager->flush();
 
+            $this->addFlash('success', 'Produit ajouté avec succès !');
+
             return $this->redirectToRoute('product_list');
         }
 
@@ -75,13 +82,38 @@ class ProductController extends AbstractController
         $locale = $request->getSession()->get('_locale', 'en');
         $localeSwitcher->setLocale($locale);
 
+        $existingImage = $product->getImage();
         $form = $this->createForm(ProductType::class, $product);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $product->checkStockStatus();
+
+            /** @var UploadedFile $imageFile */
+            $imageFile = $form->get('imageFile')->getData();
+        
+            if ($imageFile) {
+                $imageDirectory = $this->getParameter('kernel.project_dir') . '/public/assets/img/product/';
+                $newImageName = $product->getName() . '.' . $imageFile->guessExtension();
+                
+                if ($existingImage && file_exists($imageDirectory . basename($existingImage->getUrl()))) {
+                    unlink($imageDirectory . basename($existingImage->getUrl()));
+                }
+
+                $imageFile->move($imageDirectory, $newImageName);
+
+                $image = $existingImage ?? new Image();
+                $image->setUrl('/assets/img/product/' . $newImageName);
+                $image->setProduct($product);
+
+                $entityManager->persist($image);
+                $product->setImage($image);
+            }
+
             $entityManager->persist($product);
             $entityManager->flush();
+
+            $this->addFlash('info', 'Produit modifié avec succès !');
 
             return $this->redirectToRoute('product_list');
         }
@@ -95,11 +127,42 @@ class ProductController extends AbstractController
     #[Route('/product/delete/{id}', name: 'product_delete', methods: ['POST'])]
     public function delete(Request $request, Product $product, EntityManagerInterface $entityManager)
     {
-        if ($this->isCsrfTokenValid('delete'.$product->getId(), $request->request->get('_token'))) {
+        if ($this->isCsrfTokenValid('delete' . $product->getId(), $request->request->get('_token'))) {
+            $orderItems = $product->getOrderItems();
+            
+            if (!$orderItems->isEmpty()) {
+                $this->addFlash('danger', "Ce produit ne peut pas être supprimé car il fait partie d'une commande.");
+                return $this->redirectToRoute('product_list');
+            }
+            
             $entityManager->remove($product);
             $entityManager->flush();
+    
+            $this->addFlash('success', 'Produit supprimé avec succès !');
+        } else {
+            $this->addFlash('danger', 'Échec de la suppression du produit. Token CSRF invalide.');
         }
 
         return $this->redirectToRoute('product_list');
+    }
+
+    #[Route('/product/search', name: 'product_search', methods: ['GET'])]
+    public function search(Request $request, ProductRepository $productRepository): JsonResponse
+    {
+        $query = $request->query->get('q', '');
+        $products = $productRepository->searchByName($query);
+    
+        $results = [];
+        foreach ($products as $product) {
+            $results[] = [
+                'id' => $product->getId(),
+                'name' => $product->getName(),
+                'description' => $product->getDescription(),
+                'price' => $product->getPrice(),
+                'image' => $product->getImage() ? $product->getImage()->getUrl() : null,
+            ];
+        }
+    
+        return new JsonResponse($results);
     }
 }
